+++ src/main/java/com/team/buddyya/certification/controller/PhoneAuthenticationController.java
@@ -4,7 +4,9 @@ import com.team.buddyya.certification.dto.request.SendCodeRequest;
 import com.team.buddyya.certification.dto.request.VerifyCodeRequest;
 import com.team.buddyya.certification.dto.response.SendCodeResponse;
 import com.team.buddyya.certification.dto.response.VerifyCodeResponse;
+import com.team.buddyya.certification.service.MessageSendService;
 import com.team.buddyya.certification.service.PhoneAuthenticationService;
+import com.team.buddyya.certification.service.MessageSenderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,22 +20,19 @@ import org.springframework.web.bind.annotation.RestController;
 public class PhoneAuthenticationController {
 
     private final PhoneAuthenticationService phoneAuthenticationService;
+    private final MessageSendService messageSendService;
 
     @PostMapping("/send-code")
     public ResponseEntity<SendCodeResponse> sendOne(@RequestBody SendCodeRequest sendCodeRequest) {
-
-        String generatedCode = phoneAuthenticationService.sendSms(sendCodeRequest.phoneNumber());
+        String generatedCode = messageSendService.sendMessage(sendCodeRequest.phoneNumber());
         SendCodeResponse response = phoneAuthenticationService.saveCode(sendCodeRequest.phoneNumber(), generatedCode);
-
         return ResponseEntity.ok(response);
     }
 
     @PostMapping("/verify-code")
     public ResponseEntity<VerifyCodeResponse> verifyCode(@RequestBody VerifyCodeRequest verifyCodeRequest) {
-
         phoneAuthenticationService.verifyCode(verifyCodeRequest.phoneNumber(), verifyCodeRequest.code());
         VerifyCodeResponse response = phoneAuthenticationService.checkMembership(verifyCodeRequest.phoneNumber());
-
         return ResponseEntity.ok(response);
     }
 }
